Improved the regression suite's submit_tests.py's output and removed 0 as a valid test type.
In order to better parse the output put in by submit_test, added some delimiters. Also added another return type so that it would be readily apparent when there is an issue with a line in the test list file. Before, the info was only contained in the log. Unless someone noticed that their summary email didn't include a particular test, they would not have known that there was a problem in the test list file. Now a warning is raised and printed in the summaries.

Decided that users probably don't want to change a line in the test list file to have a test type of 0 in order to get tests to not run. Since commenting out is supported and generally easier, removed 0 as a valid number. Now, if 0 is seen, a syntax error is raised on that line.


git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/regression@1019 979fb63a-54b4-4514-8d67-9ab8ea1b59df