Moved the regression suite's checks for plfs and experiment_managemet config files and plfs directories to be outside of the build stage of a regression run.
In order to separate configuration checks from build checks, config file and directory checks were moved to be outside the section of code that can be skipped by using --nobuild. This way, if building is successful, but running is not due to errors in the plfs or experiment_management configuration files, rebuilding is not needed. Instead, we can now just use --nobuild to skip building and going straight to the config checks.


git-svn-id: https://plfs.svn.sourceforge.net/svnroot/plfs/regression@1022 979fb63a-54b4-4514-8d67-9ab8ea1b59df