Fixed an open without a close bug in truncate_ops.
There was a file open call without a corresponding file close call which
was making it hard (if not impossible) to unmount the plfs mount point
after the test completed. The file handle was still floating around
somewhere, so when it was time to unmount, fuse reported the device or
resource as busy. Now unmount happens successfully.