mkdir_ops.tcsh was exiting before unmounting PLFS.
All of the exit statements were left in from the original version of
mkdir_ops.tcsh. This means that if there was a problem in any of the
tests, the script exited without trying to unmount PLFS. The code has
been reworked so that this no longer happens. Plus, each test will now
be performed, no matter if the other tests failed.