Added pexec to helper scripts and updated dependencies
Since pexec isn't available outside of LANL right now, we decided to put
it into the regression suite. It already has approval for this to be
done. The version added to the regression suite is 1.4 and a comment was
added to it to reflect this version.

The documentation was updated to remove pexec as a dependency and the
descriptions of the other dependencies were clarifyed due to some
feedback.