data-sieving test now has all of the needed files.
The data-sieving test should now have all of the files that it needs. It
also should have all of the code that it needs. However, there is
currently an issue with Open MPI linking against the PLFS library that
is causing an issue when running this test's compiled binary. Once that
issue is resolved, additional testing needs to be done to make sure this
test functions as expected.