When looking for scratch spaces, non-readable directories are ignored.
rs_scratch_mount_find.sh will now not pay attention to directories that
the user doesn't have permissions to read. The find statement would
return spurious errors that python would capture and put in to the
variable. This was causing problems when python went to put that file in
to a script to run (like in cp_plfs_read).