Add some troubleshooting information to the fd_consistency test.
Since this test can be quite demanding and errors that it runs in to
aren't all PLFS errors, it would be good to make some notes about what
to check when an error occurs. For this commit, added information about
some of the resources that configure and make use and the use of mlog
with this test may be dangerous.